Tag Policy, Detailed Documentation, Postman Collections and some other quick updates

+++ app/Http/Controllers/TagController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Http\Requests\StoreTagRequest;
 use App\Http\Requests\UpdateTagRequest;
 use App\Models\Tag;
+use App\Models\Task;
 use Illuminate\Http\Request;
 
 class TagController extends Controller
@@ -14,6 +15,9 @@ class TagController extends Controller
      */
     public function index()
     {
+        // Any user can see all Tags
+        $this->authorize('viewAny', Tag::class);
+
         $tags = Tag::all();
         return response()->json($tags);
     }
@@ -34,6 +38,9 @@ class TagController extends Controller
      */
     public function update(UpdateTagRequest $request, Tag $tag)
     {
+        // Admin can update any task and user can only update their own tasks
+        $this->authorize('update', $tag);
+
         $validatedTag = $request->validated();
         $tag->update($validatedTag);
 
@@ -45,6 +52,9 @@ class TagController extends Controller
      */
     public function destroy(Tag $tag)
     {
+        // Only admin can delete the Tag
+        $this->authorize('delete', $tag);
+
         // Remove the deleted tag from the tasks
         $tag->tasks()->detach();
